* Правки по замечаниям

+++ src/parser/command_parser.hpp
@@ -9,16 +9,16 @@ enum class CommandParserState
 class CommandParser
 {
 public:
-    explicit CommandParser(size_t cmd_block_size);
+    explicit CommandParser(long cmd_block_size);
     void add_printer(Printer *printer);
     void process(const std::string &line);
     void end_parse();
 
 private:
-    size_t cmd_block_size_;
+    long cmd_block_size_;
     CommandParserState state_ = CommandParserState::BUILD_STATIC_BLOCK;
-    int nest_level_ = 0;
-    size_t cmd_count_ = 0;
+    long nest_level_ = 0;
+    long cmd_count_ = 0;
     std::vector<std::string> buffer_;
     std::vector<Printer *> printers_;
